Version 1.0: *~ Hoping to be consistent with my commit messages, each commit message will contain: which files were added, which were changed, what each file contains\what content got updated, and what is the point of this version ~* - Added main.py, claims1.csv, preprocess.py, README.txt - claims1.csv currently contains only some of the claims, and has claims that should be ignored (claims['Patent'] == 'ignore') - preprocess.py currently contains some prototypical functions that should preprocess our claims - README currently only specifies which versions of python and spacy we're using. Also, which pipeline - main.py currently processes all the claims we have but without any preprocessing...
POINT OF THE VERSION:
Well... It took time for me to find a way to use this specific model. It didn't work with newer versions of python, and it was difficult to execute it.
One might ask: why didn't you just switched to another model, one that works? Answer: mostly because I'm inexperienced and motivated, a combination that causes you to hyperfocus on one solution, while being blind to knowing when this doesn't work.