Version 1.1: *~ Hoping to be consistent with my commit messages, each commit message will contain: which files were added, which were changed, what each file contains\what content got updated, and what is the point of this version ~* - Changed main.py, claims1.csv, preprocess.py, README.txt - claims1.csv: some of the text here became more similar to the source. went only for things from patent 2. - preprocess.py fixed so that it will work with pd. also clean() became simpler. - README: added things you need to download in order for this to work. - main.py uses preprocessing now. stop_words() and lemmatize() are unnecessary.
POINT OF THE VERSION:
Well... It took time for me to find a way to use this specific model. It didn't work with newer versions of python, and it was difficult to execute it.
One might ask: why didn't you just switched to another model, one that works? Answer: mostly because I'm inexperienced and motivated, a combination that causes you to hyperfocus on one solution, while being blind to knowing when this doesn't work.